+++ NEWS.md
@@ -952,6 +952,8 @@ Deprecated or removed
      have been deprecated. Subtypes of `AbstractArray` that implement the newly introduced strided
      array interface should define their own `strides` method ([#25321]).
 
+  * `module_parent` has been deprecated in favor of `parentmodule` ([#TODO]).
+
   * `rand(t::Tuple{Vararg{Int}})` is deprecated in favor of `rand(Float64, t)` or `rand(t...)`;
     `rand(::Tuple)` will have another meaning in the future ([#25429], [#25278]).